scripts/evaluate_onnx.py
#!/usr/bin/env python3
# filename: scripts/evaluate_onnx.py
"""
PiEdge EduKit - ONNX Model Evaluator
Evaluates ONNX models and compares with PyTorch models.

Additionally (optional), can produce a simple confusion matrix image to
`reports/confusion_matrix.png` using the small training dataset layout in
`data/train/<class>/sample_*.png` when `--save-confusion-matrix` is provided.
This is intended to make verification fully green without depending on
external datasets.
"""

import os
import sys
import argparse
from pathlib import Path
from typing import List, Tuple

import numpy as np
import onnxruntime as ort
import torch
from PIL import Image
import matplotlib.pyplot as plt


def _list_images_with_labels(root: Path) -> List[Tuple[Path, int, str]]:
    images: List[Tuple[Path, int, str]] = []
    if not root.exists():
        return images
    class_dirs = [p for p in sorted(root.iterdir()) if p.is_dir()]
    for class_index, class_dir in enumerate(class_dirs):
        for img_path in sorted(class_dir.glob("*.png")):
            images.append((img_path, class_index, class_dir.name))
    return images


def _load_image_as_tensor(path: Path, size: Tuple[int, int] = (64, 64)) -> np.ndarray:
    # Load PNG and convert to CHW float32 in [0,1]
    img = Image.open(path).convert("RGB").resize(size)
    arr = np.asarray(img, dtype=np.float32) / 255.0  # HWC
    chw = np.transpose(arr, (2, 0, 1))  # CHW
    return chw


def _save_confusion_matrix_image(
    cm: np.ndarray, class_names: List[str], out_path: Path
) -> None:
    out_path.parent.mkdir(parents=True, exist_ok=True)
    fig, ax = plt.subplots(figsize=(4, 4), dpi=150)
    im = ax.imshow(cm, cmap="Blues")
    ax.set_xticks(range(len(class_names)))
    ax.set_yticks(range(len(class_names)))
    ax.set_xticklabels(class_names)
    ax.set_yticklabels(class_names)
    ax.set_xlabel("Predicted")
    ax.set_ylabel("True")
    for i in range(cm.shape[0]):
        for j in range(cm.shape[1]):
            ax.text(j, i, int(cm[i, j]), ha="center", va="center", color="black")
    fig.colorbar(im, ax=ax, fraction=0.046, pad=0.04)
    fig.tight_layout()
    fig.savefig(out_path)
    plt.close(fig)


def evaluate_onnx_model(
    onnx_path, pytorch_path=None, save_confmat: bool = False, limit: int = 0
):
    """Evaluate ONNX model and optionally compare with PyTorch model"""
    print(f"[INFO] Evaluating ONNX model: {onnx_path}")

    if not Path(onnx_path).exists():
        print(f"[ERROR] ONNX model not found: {onnx_path}")
        return False

    try:
        # Load ONNX model
        session = ort.InferenceSession(onnx_path)
        input_name = session.get_inputs()[0].name
        output_name = session.get_outputs()[0].name

        print(f"[OK] ONNX model loaded successfully")
        print(f"   Input: {input_name} {session.get_inputs()[0].shape}")
        print(f"   Output: {output_name} {session.get_outputs()[0].shape}")

        # Test with dummy data
        dummy_input = np.random.randn(1, 3, 64, 64).astype(np.float32)

        # ONNX inference
        onnx_outputs = session.run([output_name], {input_name: dummy_input})
        onnx_output = onnx_outputs[0]

        print(
            f"[OK] ONNX inference successful: {onnx_output.shape} {onnx_output.dtype}"
        )

        # Compare with PyTorch if available
        if pytorch_path and Path(pytorch_path).exists():
            print(f"[INFO] Comparing with PyTorch model: {pytorch_path}")

            try:
                # Load PyTorch model (could be state_dict or full model)
                pytorch_data = torch.load(pytorch_path, map_location="cpu")

                # Check if it's a state_dict or full model
                if isinstance(pytorch_data, dict) and "state_dict" in pytorch_data:
                    # It's a checkpoint with state_dict
                    print("[INFO] Loading from checkpoint with state_dict")
                    # For now, skip comparison as we need the model architecture
                    print(
                        "[WARN] Cannot compare with state_dict checkpoint - skipping comparison"
                    )
                elif hasattr(pytorch_data, "eval"):
                    # It's a full model
                    pytorch_model = pytorch_data
                    pytorch_model.eval()

                    # PyTorch inference
                    with torch.no_grad():
                        pytorch_input = torch.from_numpy(dummy_input)
                        pytorch_output = pytorch_model(pytorch_input).numpy()

                    # Compare outputs
                    diff = np.abs(onnx_output - pytorch_output).max()
                    print(f"[INFO] Max difference: {diff:.6f}")

                    if diff < 1e-5:
                        print("[OK] ONNX and PyTorch outputs match!")
                    else:
                        print("[WARN] ONNX and PyTorch outputs differ significantly")
                else:
                    print("[WARN] Unknown PyTorch model format - skipping comparison")
            except Exception as e:
                print(f"[WARN] Error loading PyTorch model: {e} - skipping comparison")

        # Optionally compute a tiny confusion matrix using data/train
        if save_confmat:
            data_root = Path("data") / "train"
            items = _list_images_with_labels(data_root)
            if limit > 0:
                items = items[:limit]
            if items:
                num_classes = len(sorted({name for _, _, name in items}))
                class_names = sorted({name for _, _, name in items})
                name_to_index = {name: idx for idx, name in enumerate(class_names)}
                cm = np.zeros((num_classes, num_classes), dtype=np.int32)
                for img_path, true_index_unused, class_name in items:
                    true_index = name_to_index[class_name]
                    x = _load_image_as_tensor(img_path)[None, ...]  # 1x3x64x64
                    y = session.run([output_name], {input_name: x.astype(np.float32)})[
                        0
                    ]
                    pred_index = (
                        int(np.argmax(y, axis=1)[0])
                        if y.ndim == 2
                        else int(np.argmax(y))
                    )
                    if 0 <= true_index < num_classes and 0 <= pred_index < num_classes:
                        cm[true_index, pred_index] += 1
                _save_confusion_matrix_image(
                    cm, class_names, Path("reports") / "confusion_matrix.png"
                )
                print("[OK] Saved confusion matrix to reports/confusion_matrix.png")
            else:
                print(
                    "[WARN] No images found under data/train – skipping confusion matrix save"
                )

        return True

    except Exception as e:
        print(f"[ERROR] Error evaluating ONNX model: {e}")
        return False


def main():
    parser = argparse.ArgumentParser(description="Evaluate ONNX model")
    parser.add_argument("--model", required=True, help="Path to ONNX model")
    parser.add_argument(
        "--fakedata", action="store_true", help="Use fake data for evaluation"
    )
    parser.add_argument("--limit", type=int, default=16, help="Limit number of samples")
    parser.add_argument(
        "--save-confusion-matrix",
        action="store_true",
        help="Save reports/confusion_matrix.png using images from data/train",
    )
    args = parser.parse_args()

    print("[INFO] PiEdge EduKit - ONNX Model Evaluator")
    print("=" * 50)

    # Check for models
    onnx_model = Path(args.model)
    if not onnx_model.exists():
        print(f"[ERROR] ONNX model not found: {onnx_model}")
        sys.exit(1)

    # Look for PyTorch model in same directory
    pytorch_model = onnx_model.parent / "model_best.pth"
    if not pytorch_model.exists():
        pytorch_model = None

    print(f"[INFO] Using ONNX model: {onnx_model}")
    if pytorch_model:
        print(f"[INFO] Using PyTorch model: {pytorch_model}")
    else:
        print("[INFO] No PyTorch model found for comparison")

    # Evaluate model
    if evaluate_onnx_model(
        onnx_model,
        pytorch_model,
        save_confmat=args.save_confusion_matrix,
        limit=max(0, int(args.limit)),
    ):
        print("\n[SUCCESS] ONNX model evaluation completed!")
    else:
        print("\n[WARN] ONNX model evaluation failed!")
        sys.exit(1)


if __name__ == "__main__":
    main()
